// kit-plugin-metaplex/src/generated/mpl_token_metadata-client/accounts/metadata.ts
export type Metadata = {
  key: Key;
  updateAuthority: Address;
  mint: Address;
  data: Data;
  primarySaleHappened: boolean;
  isMutable: boolean;
  editionNonce: Option<number>;
  tokenStandard: Option<TokenStandard>;
  collection: Option<Collection>;
  uses: Option<Uses>;
  collectionDetails: Option<CollectionDetails>;
  programmableConfig: Option<ProgrammableConfig>;
};